refactor: 리팩토링
- Home 페이지의 컴포넌트 리팩토링
- 검색 기능 추가를 위한 폴더명 및 상태 관리 변경
